pyproject.toml
[project]
name = "fast-zero-async"
version = "0.1.0"
description = ""
authors = [
    {name = "ClaudioFritzen",email = "sdfritzen96@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14, <4.0"
dependencies = [
    "fastapi[standard] (>=0.123.8,<0.124.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "pwdlib[argon2] (>=0.3.0,<0.4.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "tzdata (>=2025.2,<2026.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.14.8,<0.15.0)",
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "taskipy (>=1.14.1,<2.0.0)"
]

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings ' # to ignore warnings during test runsc

[tool.taskipy.tasks]
lint = "ruff check ."
format = "ruff format ."
run = 'fastapi dev fast_zero_async/app.py'
test = 'pytest -s -x --cov=fast_zero_async -vv'
pre_format = 'ruff check --fix'
pre_test = 'task lint'
post_test = 'coverage html'
run_ip ='fastapi dev fast_zero_async/app.py --host 192.168.1.66'
uvicorn = 'uvicorn fast_zero_async.app:app --reload --port 8001'
